Move election package generation function into its own file
